BGDIINF_SB-1478: field type of Asset has to be required and validated
Putting all validators into a sepperate file

clean() media_type without space

repaired tests as media_type without space
