+++ app/stac_api/validators.py
@@ -1,0 +1,196 @@
+import logging
+import re
+from datetime import datetime
+
+from django.contrib.gis.geos import GEOSGeometry
+from django.core.exceptions import ValidationError
+from django.utils.translation import gettext_lazy as _
+from stac_api.utils import fromisoformat
+
+logger = logging.getLogger(__name__)
+
+MEDIA_TYPES = [
+    ('application/x.asc+zip', 'Zipped ASCII (.asc)', '.zip'),
+    ('application/x.asc-grid+zip', 'Zipped ASCII-GRID (.xyz)', '.zip'),
+    ('application/x.e00+zip 	Zipped', 'e00', '.zip'),
+    ('image/tiff; application=geotiff', 'GeoTIFF', '.tiff or .tif'),
+    ('application/x.geotiff+zip', 'Zipped GeoTIFF', '.zip'),
+    ('application/x.tiff+zip', 'Zipped TIFF', '.zip'),
+    ('application/x.png+zip', 'Zipped PNG', '.zip'),
+    ('application/x.jpeg+zip', 'Zipped JPEG', '.zip'),
+    ('application/vnd.google-earth.kml+xml', 'KML', '.kml'),
+    ('application/vnd.google-earth.kmz', 'Zipped KML', '.kmz'),
+    ('application/x.dxf+zip', 'Zipped DXF', '.zip'),
+    ('application/gml+xml', 'GML', '.gml or .xml'),
+    ('application/x.gml+zip', 'Zipped GML', '.zip'),
+    ('application/vnd.las', 'LIDAR', '.las'),
+    ('application/vnd.laszip', 'Zipped LIDAR', '.laz or .zip'),
+    ('application/x.shapefile+zip', 'Zipped Shapefile', '.zip'),
+    ('application/x.filegdb+zip', 'Zipped File Geodatabase', '.zip'),
+    ('application/x.ms-access+zip', 'Zipped Personal Geodatabase', '.zip'),
+    ('application/x.ms-excel+zip', 'Zipped Excel', '.zip'),
+    ('application/x.tab+zip', 'Zipped Mapinfo-TAB', '.zip'),
+    ('application/x.tab-raster+zip', 'Zipped Mapinfo-Raster-TAB', '.zip'),
+    ('application/x.csv+zip', 'Zipped CSV', '.zip'),
+    ('text/csv', 'CSV', '.csv'),
+    ('application/geopackage+sqlite3', 'Geopackage', '.gpkg'),
+    ('application/x.geopackage+zip', 'Zipped Geopackage', '.zip'),
+    ('application/geo+json', 'GeoJSON', '.json or .geojson'),
+    ('application/x.geojson+zip', 'Zipped GeoJSON', '.zip'),
+    ('application/x.interlis;version=2.3', 'Interlis 2', '.xtf or .xml'),
+    ('application/x.interlis+zip;version=2.3', 'Zipped XTF (2.3)', '.zip'),
+    ('application/x.interlis;version=1', 'Interlis 1', '.itf'),
+    ('application/x.interlis+zip;version=1', 'Zipped ITF', '.zip'),
+    (
+        'image/tiff;application=geotiff;profile=cloud-optimize',
+        'Cloud Optimized GeoTIFF (COG)',
+        '.tiff or .tif'
+    ),
+    ('application/pdf', 'PDF', '.pdf'),
+    ('application/x.pdf+zip', 'Zipped PDF', '.zip'),
+    ('application/json', 'JSON', '.json'),
+    ('application/x.json+zip', 'Zipped JSON', '.zip'),
+    ('application/x-netcdf', 'NetCDF', '.nc'),
+    ('application/x.netcdf+zip', 'Zipped NetCDF', '.zip'),
+    ('application/xml', 'XML', '.xml'),
+    ('application/x.xml+zip', 'Zipped XML', '.zip'),
+    ('application/vnd.mapbox-vector-tile', 'mbtiles', '???'),
+    ('text/plain', 'Text', '.txt'),
+    ('text/x.plain+zip', 'Zipped text', '.zip')
+]
+MEDIA_TYPES_MIMES = [x[0] for x in iter(MEDIA_TYPES)]
+
+
+def validate_media_types(value):
+    # replace - as the mime type is sometimes with space
+    if value.replace(' ', '') not in MEDIA_TYPES_MIMES:
+        message = f"The media type defined must be in {MEDIA_TYPES_MIMES}"
+        logger.error(message)
+        raise ValidationError(_(message))
+
+
+def validate_name(name):
+    '''Validate name used in URL
+    '''
+    if not re.match(r'^[0-9a-z-_.]+$', name):
+        logger.error('Invalid name %s, only the following characters are allowed: 0-9a-z-_.', name)
+        raise ValidationError(
+            _('Invalid name, only the following characters are allowed: 0-9a-z-_.')
+        )
+
+
+def validate_geoadmin_variant(variant):
+    '''Validate geoadmin:variant, it should not have special characters'''
+    if not re.match('^[a-zA-Z0-9]+$', variant):
+        logger.error(
+            "Invalid geoadmin:variant property %s, special characters not allowed", variant
+        )
+        raise ValidationError(_("Invalid geoadmin:variant, special characters not allowed"))
+
+
+def validate_link_rel(value):
+    invalid_rel = [
+        'self',
+        'root',
+        'parent',
+        'items',
+        'collection',
+        'service-desc',
+        'service-doc',
+        'search',
+        'conformance'
+    ]
+    if value in invalid_rel:
+        logger.error("Link rel attribute %s is not allowed, it is a reserved attribute", value)
+        raise ValidationError(_(f'Invalid rel attribute, must not be in {invalid_rel}'))
+
+
+def validate_geometry(geometry):
+    '''
+    A validator function that ensures, that only valid
+    geometries are stored.
+    Args:
+         geometry: The geometry that will be validated
+
+    Returns:
+        The geometry, when tested valid
+
+    Raises:
+        ValidateionError: About that the geometry is not valid
+    '''
+    geos_geometry = GEOSGeometry(geometry)
+    if geos_geometry.empty:
+        message = "The geometry is empty: %s" % geos_geometry.wkt
+        logger.error(message)
+        raise ValidationError(_(message))
+    if not geos_geometry.valid:
+        message = "The geometry is not valid: %s" % geos_geometry.valid_reason
+        logger.error(message)
+        raise ValidationError(_(message))
+    return geometry
+
+
+def validate_item_properties_datetimes_dependencies(
+    properties_datetime, properties_start_datetime, properties_end_datetime
+):
+    '''
+    Validate the dependencies between the Item datetimes properties
+	This makes sure that either only the properties.datetime is set or
+	both properties.start_datetime and properties.end_datetime
+	Raises:
+		django.core.exceptions.ValidationError
+    '''
+    try:
+        if not isinstance(properties_datetime, datetime) and properties_datetime is not None:
+            properties_datetime = fromisoformat(properties_datetime)
+        if (
+            not isinstance(properties_start_datetime, datetime) and
+            properties_start_datetime is not None
+        ):
+            properties_start_datetime = fromisoformat(properties_start_datetime)
+        if (
+            not isinstance(properties_end_datetime, datetime) and
+            properties_end_datetime is not None
+        ):
+            properties_end_datetime = fromisoformat(properties_end_datetime)
+    except ValueError as error:
+        logger.error("Invalid datetime string %s", error)
+        raise ValidationError(f'Invalid datetime string {error}') from error
+
+    if properties_datetime is not None:
+        if (properties_start_datetime is not None or properties_end_datetime is not None):
+            message = 'Cannot provide together property datetime with datetime range ' \
+                '(start_datetime, end_datetime)'
+            logger.error(message)
+            raise ValidationError(_(message))
+    else:
+        if properties_end_datetime is None:
+            message = "Property end_datetime can't be null when no property datetime is given"
+            logger.error(message)
+            raise ValidationError(_(message))
+        if properties_start_datetime is None:
+            message = "Property start_datetime can't be null when no property datetime is given"
+            logger.error(message)
+            raise ValidationError(_(message))
+
+    if properties_datetime is None:
+        if properties_end_datetime < properties_start_datetime:
+            message = "Property end_datetime can't refer to a date earlier than property "\
+            "start_datetime"
+            raise ValidationError(_(message))
+
+
+def validate_item_properties_datetimes(
+    properties_datetime, properties_start_datetime, properties_end_datetime, partial=False
+):
+    '''
+    Validate datetime values in the properties Item attributes
+    '''
+    if not partial:
+        # Do not validate dependencies in partial update, leave it to the validation to the model
+        # instance.
+        validate_item_properties_datetimes_dependencies(
+            properties_datetime,
+            properties_start_datetime,
+            properties_end_datetime,
+        )
